requirements.txt
Django==1.8.2
django-bootstrap3==5.4.0
django-debug-toolbar==1.3.0
django-extensions==1.5.5
fake-factory==0.5.2
ipython==3.1.0
psycopg2==2.6.1
Werkzeug==0.10.4
autoenv==1.0.0
beautifulsoup4==4.3.2
certifi==2015.4.28
Cython==0.22
Django==1.8.2
django-bootstrap3==5.4.0
django-chartit==0.1
django-debug-toolbar==1.3.0
django-extensions==1.5.5
djangorestframework==3.1.3
djangorestframework-jwt==1.6.0
entities==1.0.0
epc==0.0.5
fake-factory==0.5.2
folium==0.1.4
gnureadline==6.3.3
hashids==0.8.4
httpie==0.9.2
ipython==3.1.0
jedi==0.9.0
Jinja2==2.7.3
jsonschema==2.4.0
MarkupSafe==0.23
matplotlib==1.4.3
mistune==0.5.1
nose==1.3.1
numpy==1.9.2
pandas==0.16.1
pep8==1.6.2
picamera==1.10
Pillow==2.8.1
psycopg2==2.6.1
ptyprocess==0.4
py==1.4.27
py-postgresql==1.0.3
pyglet==1.2.2
Pygments==2.0.2
pygoogle==0.6
PyJWT==1.3.0
pyparsing==2.0.3
pytest==2.7.1
python-dateutil==2.4.2
pytz==2015.4
pyzmq==14.6.0
requests==2.7.0
scikit-learn==0.16.1
scipy==0.15.1
seaborn==0.5.1
sexpdata==0.0.3
six==1.9.0
sqlparse==0.1.15
terminado==0.5
tkinter==0.0.0
tornado==4.2
virtualenv==12.1.1
Werkzeug==0.10.4
